+++ hutool-extra/src/main/java/cn/hutool/extra/ssh/JschUtil.java
@@ -31,8 +31,6 @@ public class JschUtil {
 
 	/** 本地端口生成器 */
 	private static final LocalPortGenerater portGenerater = new LocalPortGenerater(10000);
-	/** 锁 */
-	private static final Object lock = new Object();
 
 	/**
 	 * 生成一个本地端口，用于远程端口映射
@@ -53,18 +51,7 @@ public class JschUtil {
 	 * @return SSH会话
 	 */
 	public static Session getSession(String sshHost, int sshPort, String sshUser, String sshPass) {
-		final String key = StrUtil.format("{}@{}:{}", sshUser, sshHost, sshPort);
-		Session session = JschSessionPool.INSTANCE.get(key);
-		if (null == session) {
-			synchronized (lock) {
-				session = JschSessionPool.INSTANCE.get(key);
-				if (null == session || false == session.isConnected()) {
-					session = openSession(sshHost, sshPort, sshUser, sshPass);
-					JschSessionPool.INSTANCE.put(key, session);
-				}
-			}
-		}
-		return session;
+		return JschSessionPool.INSTANCE.getSession(sshHost, sshPort, sshUser, sshPass);
 	}
 
 	/**
@@ -108,7 +95,7 @@ public class JschUtil {
 			try {
 				session.setPortForwardingL(localPort, remoteHost, remotePort);
 			} catch (JSchException e) {
-				throw new JschRuntimeException("From [" + remoteHost + "] Mapping to [" + localPort + "] error！", e);
+				throw new JschRuntimeException(e, "From [{}] mapping to [{}] error！", remoteHost, localPort);
 			}
 			return true;
 		}
@@ -259,6 +246,7 @@ public class JschUtil {
 		if (session != null && session.isConnected()) {
 			session.disconnect();
 		}
+		JschSessionPool.INSTANCE.remove(session);
 	}
 
 	/**